+++ src/menu/Operation.java
@@ -1,0 +1,13 @@
+package menu;
+
+public interface Operation{
+
+  public void execute();
+
+}
+
+
+
+
+
+
