+++ src/menu/Operation.~ava
@@ -1,0 +1,13 @@
+package wwa;
+
+public interface Operation{
+
+  public void execute();
+
+}
+
+
+
+
+
+
