+++ src/menu/OperationExecutor.java
@@ -1,0 +1,13 @@
+package menu;
+import java.util.ArrayList;
+
+public class OperationExecutor {
+     
+    static final ArrayList <Operation> Operations = new ArrayList<>();
+     
+     static void executeOperation(Operation operation) {
+        Operations.add(operation);
+        operation.execute();
+    }
+}
+
