+++ src/menu/StartGame.~ava
@@ -1,0 +1,20 @@
+package menu;
+
+public class StartGame implements Operation {
+ 
+    private Option option;
+     
+    // constructors
+    public StartGame() {
+      //option musste noch initialisiert werden
+      option = new Option();
+    }
+     
+    @Override
+    public void execute() {
+        
+        option.start();
+        
+    }
+}
+
