--- src/wwa/ExitGame.~ava
@@ -1,12 +1,0 @@
-package wwa;
-
-
-public class ExitGame implements Operation {
-     
-    // same field and constructor as above
-         
-    @Override
-    public void execute() {
-        return option.exit();
-    }
-}
