--- src/wwa/LoadGame.~ava
@@ -1,11 +1,0 @@
-package wwa;
-
-public class LoadGame implements Operation {
-     
-    private Option option;
-         
-    @Override
-    public void execute() {
-        option.load();
-    }
-}
