--- src/wwa/Operation.java
@@ -1,13 +1,0 @@
-package wwa;
-
-public interface Operation{
-
-  public void execute();
-
-}
-
-
-
-
-
-
