--- src/wwa/Operation.~ava
@@ -1,13 +1,0 @@
-package wwa;
-
-interface Operation{
-
-  public void execute();
-
-}
-
-
-
-
-
-
