--- src/wwa/Option.java
@@ -1,26 +1,0 @@
-package wwa;
-
-
-public class Option {
-     
-    
-     
-    // constructor
-     
-    public String start() {
-       System.out.print("The game is loading...");    
-       return "The game is loading...";
-    }
-     
-    public String load() {  
-        System.out.print("The game is loading..."); 
-        return "Load savestate ";
-    }
-  
-   public String exit() {
-        System.out.print("Exiting the game."); 
-        return "Exit Game ";
-    }
-     
-                                                                                            
-}
