--- src/wwa/Option.~ava
@@ -1,23 +1,0 @@
-package wwa;
-
-
-public class Option {
-     
-    
-     
-    // constructor
-     
-    public String start() {
-       return "Das Game wird gestartet...";
-    }
-     
-    public String load() {  
-        return "Load savestate ";
-    }
-  
-   public String exit() {
-        return "Exit Game ";
-    }
-     
-                                                                                            
-}
