--- src/wwa/StartGame.java
@@ -1,20 +1,0 @@
-package wwa;
-
-public class StartGame implements Operation {
- 
-    private Option option;
-     
-    // constructors
-    public StartGame() {
-    	//option musste noch initialisiert werden
-    	option = new Option();
-    }
-     
-    @Override
-    public void execute() {
-        
-        option.start();
-        
-    }
-}
-
