--- src/wwa/StartGame.~ava
@@ -1,14 +1,0 @@
-package wwa;
-
-public class StartGame implements Operation {
- 
-    private Option option;
-     
-    // constructors
-     
-    @Override
-    public void execute() {
-        option.start();
-    }
-}
-
